Immediately apply pref changes
Interacting with pref controls will have an immediate effect, without
having to click a save button, which is therefore removed. Instead,
there is a new undo button, which revert all the current changes to the
state when the pref page was opened.

The "undo" and "reset" buttons are more intelligent: they are disabled
when they would have no effect.

Fixes #4